# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: COGS (Cost of Goods Sold) Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/services:
    get:
      tags: [Services]
      operationId: listProjectServices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectService'
    post:
      tags: [Services]
      operationId: createProjectService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectServiceBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectService'
  /services/{id}:
    patch:
      tags: [Services]
      operationId: updateProjectService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectServiceBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectService'
    delete:
      tags: [Services]
      operationId: deleteProjectService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /reports/projects/{id}/summary:
    get:
      tags: [Reports]
      operationId: getProjectSummary
      summary: Per-project summary report (requires summary access)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummaryReport'
        '403':
          description: Forbidden — user lacks summary access for this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /reports/aggregate:
    get:
      tags: [Reports]
      operationId: getAggregateReport
      summary: Cross-project aggregate breakdown by service
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: projectId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateReport'
  /reports/trends:
    get:
      tags: [Reports]
      operationId: getTrendsReport
      summary: Daily trend series of total cost and cost per manday
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: projectId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsReport'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Project:
      type: object
      required: [id, name, location, contractStart, contractEnd, createdAt, isAdminOwned, currentUserCanViewSummary, currentUserCanEditEntries]
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: string }
        contractStart: { type: string, format: date }
        contractEnd: { type: string, format: date }
        notes:
          type: [string, 'null']
        createdAt: { type: string, format: date-time }
        isAdminOwned: { type: boolean }
        currentUserCanViewSummary: { type: boolean }
        currentUserCanEditEntries: { type: boolean }
...
    ProjectService:
      type: object
      required: [id, projectId, name, kind, sortOrder]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        kind:
          type: string
          enum: [food, standard]
        sortOrder: { type: integer }
    CreateProjectServiceBody:
      type: object
      required: [name, kind]
      properties:
        name: { type: string, minLength: 1 }
        kind:
          type: string
          enum: [food, standard]
        sortOrder: { type: integer }
    UpdateProjectServiceBody:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        kind:
          type: string
          enum: [food, standard]
        sortOrder: { type: integer }
...
    DailyEntrySummary:
      type: object
      required: [id, projectId, projectName, entryDate, location, totalMandays, totalCost, costPerManday]
      properties:
        id: { type: string }
        projectId: { type: string }
        projectName: { type: string }
        entryDate: { type: string, format: date }
        location: { type: string }
        totalMandays: { type: number }
        totalCost: { type: number }
        costPerManday: { type: number }
        notes:
          type: [string, 'null']
...
    DashboardKpi:
      type: object
      required: [totalCost, totalMandays, costPerManday, entryCount]
      properties:
        totalCost: { type: number }
        totalMandays: { type: number }
        costPerManday: { type: number }
        entryCount: { type: integer }
    ServiceTotal:
      type: object
      required: [serviceName, kind, totalCost, totalMandayContribution]
      properties:
        serviceName: { type: string }
        kind:
          type: string
          enum: [food, standard]
        totalCost: { type: number }
        totalMandayContribution: { type: number }
    ProjectTotal:
      type: object
      required: [projectId, projectName, location, totalCost, totalMandays, costPerManday]
      properties:
        projectId: { type: string }
        projectName: { type: string }
        location: { type: string }
        totalCost: { type: number }
        totalMandays: { type: number }
        costPerManday: { type: number }
...
    ProjectSummaryReport:
      type: object
      required: [project, range, kpi, serviceBreakdown, dailyEntries]
      properties:
        project:
          $ref: '#/components/schemas/Project'
        range:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: date }
            to: { type: string, format: date }
        kpi:
          $ref: '#/components/schemas/DashboardKpi'
        serviceBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/ServiceTotal'
        dailyEntries:
          type: array
          items:
            $ref: '#/components/schemas/DailyEntrySummary'
    AggregateReport:
      type: object
      required: [range, kpi, serviceBreakdown, projectBreakdown]
      properties:
        range:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: date }
            to: { type: string, format: date }
        kpi:
          $ref: '#/components/schemas/DashboardKpi'
        serviceBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/ServiceTotal'
        projectBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/ProjectTotal'
    TrendPoint:
      type: object
      required: [date, totalCost, totalMandays, costPerManday]
      properties:
        date: { type: string, format: date }
        totalCost: { type: number }
        totalMandays: { type: number }
        costPerManday: { type: number }
    TrendsReport:
      type: object
      required: [range, points]
      properties:
        range:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: date }
            to: { type: string, format: date }
        points:
          type: array
          items:
            $ref: '#/components/schemas/TrendPoint'