# lib/api-zod/src/generated/types/createProjectServiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COGS (Cost of Goods Sold) Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectServiceBodyKind } from "./createProjectServiceBodyKind";

export interface CreateProjectServiceBody {
  /** @minLength 1 */
  name: string;
  kind: CreateProjectServiceBodyKind;
  sortOrder?: number;
}