# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COGS (Cost of Goods Sold) Management API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  date: Date;
  totalCost: number;
  totalMandays: number;
  costPerManday: number;
}