# lib/api-zod/src/generated/types/trendsReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COGS (Cost of Goods Sold) Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TrendPoint } from "./trendPoint";
import type { TrendsReportRange } from "./trendsReportRange";

export interface TrendsReport {
  range: TrendsReportRange;
  points: TrendPoint[];
}