# lib/api-zod/src/generated/types/updateProjectServiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COGS (Cost of Goods Sold) Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectServiceBodyKind } from "./updateProjectServiceBodyKind";

export interface UpdateProjectServiceBody {
  /** @minLength 1 */
  name?: string;
  kind?: UpdateProjectServiceBodyKind;
  sortOrder?: number;
}